% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{\%awr\%}
\alias{\%awr\%}
\title{Add with rollback}
\usage{
d \%awr\% t
}
\arguments{
\item{d}{A vector of dates.}

\item{t}{A vector of duration objects.}
}
\value{
The vector of dates obtained by adding t to d. Length of d and t must
be either the same size or size 1.
}
\description{
Custom version of the \code{add_with_rollback} function from the lubridate
package.
}
\examples{
library(lubridate)
dmy("29/01/2001") \%awr\% days(1) \%awr\% months(1)
dmy("29/01/2001") \%awr\% months(1) \%awr\% days(1)
dmy("29/02/2000") \%awr\% years(1)

}
