% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linkplots-package.R
\docType{data}
\name{immo}
\alias{immo}
\title{Immobilier}
\format{
Many Variables Regarding Real Estates Transactions
}
\usage{
immo
}
\description{
Dataset from the Kaggle Competition Website
}
\keyword{datasets}
