% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{lab_numb}
\alias{lab_numb}
\title{Labels for numbers}
\usage{
lab_numb(accuracy = NULL)
}
\arguments{
\item{accuracy}{A number to round to. Use (e.g.) \code{0.01} to show 2 decimal
places of precision. If \code{NULL}, the default, uses a heuristic that should
ensure breaks have the minimum number of digits needed to show the
difference between adjacent values.

Applied to rescaled data.}
}
\description{
Labels for numbers
}
