% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{lkp_2d_contour}
\alias{lkp_2d_contour}
\title{2D Contourplot}
\usage{
lkp_2d_contour(
  df,
  x,
  y,
  fill,
  f1,
  f2,
  breaks_fill,
  names_breaks_fill,
  n_breaks_fill = 12,
  label_fill = lab_numb,
  accuracy_fill = NULL,
  show_numbers,
  label_x,
  label_y,
  accuracy_x = NULL,
  accuracy_y = NULL,
  n_breaks_x = NULL,
  n_breaks_y = NULL,
  facet_type = "wrap",
  scales = "fixed",
  labeller = label_verbose,
  nrow = NULL,
  ncol = NULL
)
}
\arguments{
\item{df}{A data.frame containing the data to represent}

\item{x}{First dimension}

\item{y}{Second dimension}

\item{fill}{Variable of interest}

\item{f1}{First optional variable for facet display}

\item{f2}{Second optional variable for facet display}

\item{breaks_fill}{Breaks to use for the variable of interest. Optional}

\item{names_breaks_fill}{Names of breaks to use for the variable of interest. Optional}

\item{n_breaks_fill}{Number of breaks to use for the variable of interest.}

\item{label_fill}{Labeller to use for the variable of interest}

\item{accuracy_fill}{Accuracy to use in the display of the variable of
interest}

\item{show_numbers}{Should numbers be printed on top of the bars ? Optional}

\item{label_x}{Labeller for the x axis. Optional}

\item{label_y}{Labeller for the y axis. Optional}

\item{accuracy_x}{Accuracy for the y axis. Optional}

\item{accuracy_y}{Accuracy for the y axis. Optional}

\item{n_breaks_x}{Desired number of breaks for the x axis. Optional}

\item{n_breaks_y}{Desired number of breaks for the y axis. Optional}

\item{facet_type}{Should \code{facet_wrap} or \code{facet_grid} be used for
facetting ?}

\item{scales}{Should scales be fixed (\code{"fixed"}, the default),
free (\code{"free"}), or free in one dimension (\code{"free_x"},
\code{"free_y"})?}

\item{labeller}{A function that takes one data frame of labels and
returns a list or data frame of character vectors. Each input
column corresponds to one factor. Thus there will be more than
one with \code{vars(cyl, am)}. Each output
column gets displayed as one separate line in the strip
label. This function should inherit from the "labeller" S3 class
for compatibility with \code{\link[ggplot2:labeller]{labeller()}}. You can use different labeling
functions for different kind of labels, for example use \code{\link[ggplot2:labellers]{label_parsed()}} for
formatting facet labels. \code{\link[ggplot2:labellers]{label_value()}} is used by default,
check it for more details and pointers to other options.}

\item{nrow}{Number of rows and columns.}

\item{ncol}{Number of rows and columns.}
}
\value{
A ggplot object representing a contourplot.
}
\description{
Plot a coloured contourplot of the 2 dimensional dataset using the specified
breaks or number of breaks
}
