% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{lkp_density}
\alias{lkp_density}
\title{Density Plot}
\usage{
lkp_density(
  df,
  x,
  f1,
  f2,
  fill,
  ...,
  label_x,
  label_y,
  accuracy_x = NULL,
  accuracy_y = NULL,
  n_breaks_x = NULL,
  n_breaks_y = NULL,
  facet_type = "wrap",
  scales = "fixed",
  nrow = NULL,
  ncol = NULL
)
}
\arguments{
\item{df}{A data.frame containing the data to represent}

\item{x}{The x axis variable for the histogram}

\item{f1}{First optional variable for facet display}

\item{f2}{Second optional variable for facet display}

\item{fill}{A variable for which to show the breakdown in each histogram bar}

\item{...}{Additional parameters passed to the \code{"ll_density"} function}

\item{label_x}{Labeller for the x axis. Optional}

\item{label_y}{Labeller for the y axis. Optional}

\item{accuracy_x}{Accuracy for the y axis. Optional}

\item{accuracy_y}{Accuracy for the y axis. Optional}

\item{n_breaks_x}{Desired number of breaks for the x axis. Optional}

\item{n_breaks_y}{Desired number of breaks for the y axis. Optional}

\item{facet_type}{Should \code{facet_wrap} or \code{facet_grid} be used for
facetting ?}

\item{scales}{Should scales be fixed (\code{"fixed"}, the default),
free (\code{"free"}), or free in one dimension (\code{"free_x"},
\code{"free_y"})?}

\item{nrow}{Number of rows and columns.}

\item{ncol}{Number of rows and columns.}
}
\value{
A ggplot object representing a density plot with one or several
facets.
}
\description{
Density Plot
}
