% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{lkp_hist_date}
\alias{lkp_hist_date}
\title{Histogram Over Date Range}
\usage{
lkp_hist_date(
  df,
  x,
  f1,
  f2,
  fill,
  wt = NULL,
  bw = "y",
  what = "number",
  show_numbers,
  label_text,
  accuracy_text = NULL,
  position_text,
  vjust_text = -0.25,
  hjust_text = 0.5,
  label_x,
  label_y,
  accuracy_x = NULL,
  accuracy_y = NULL,
  n_breaks_x = NULL,
  n_breaks_y = NULL,
  facet_type = "wrap",
  scales = "fixed",
  labeller = label_verbose,
  nrow = NULL,
  ncol = NULL,
  ...
)
}
\arguments{
\item{df}{A data.frame containing the data to represent}

\item{x}{Variable representing a date}

\item{f1}{First optional variable for facet display}

\item{f2}{Second optional variable for facet display}

\item{fill}{A variable for which to show the breakdown in each histogram bar}

\item{wt}{Variable to use as weight}

\item{bw}{Which bandwidth should be use ? Default is \code{"y"} for one year,
other possibilities include \code{"m"} for one month and \code{"d"} for one
day.}

\item{what}{If equal to \code{"number"}, the default, counts are printed. If
equal to \code{"freq"}, overall frequency is printed instead. If equal to
\code{"subfreq"}, frequency among this group is printed instead.}

\item{show_numbers}{Should numbers be printed on top of the bars ? Optional}

\item{label_text}{Labeller for the printed numbers. Optional}

\item{accuracy_text}{Accuracy for the printed numbers. Optional}

\item{position_text}{Should the text position use the \code{vjust_text} and
\code{hjust_text} arguments or should it be placed at half the bar's height
?}

\item{vjust_text}{Vertical adjustement for text}

\item{hjust_text}{Horizontal adjustement for text}

\item{label_x}{Labeller for the x axis. Optional}

\item{label_y}{Labeller for the y axis. Optional}

\item{accuracy_x}{Accuracy for the y axis. Optional}

\item{accuracy_y}{Accuracy for the y axis. Optional}

\item{n_breaks_x}{Desired number of breaks for the x axis. Optional}

\item{n_breaks_y}{Desired number of breaks for the y axis. Optional}

\item{facet_type}{Should \code{facet_wrap} or \code{facet_grid} be used for
facetting ?}

\item{scales}{Should scales be fixed (\code{"fixed"}, the default),
free (\code{"free"}), or free in one dimension (\code{"free_x"},
\code{"free_y"})?}

\item{labeller}{A function that takes one data frame of labels and
returns a list or data frame of character vectors. Each input
column corresponds to one factor. Thus there will be more than
one with \code{vars(cyl, am)}. Each output
column gets displayed as one separate line in the strip
label. This function should inherit from the "labeller" S3 class
for compatibility with \code{\link[ggplot2:labeller]{labeller()}}. You can use different labeling
functions for different kind of labels, for example use \code{\link[ggplot2:labellers]{label_parsed()}} for
formatting facet labels. \code{\link[ggplot2:labellers]{label_value()}} is used by default,
check it for more details and pointers to other options.}

\item{nrow}{Number of rows and columns.}

\item{ncol}{Number of rows and columns.}

\item{...}{Additional parameters to be passed to \code{ll_bar}}
}
\value{
A ggplot object representing an histogram with one or several facets
}
\description{
Histogram Over Date Range
}
