% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{lkp_pyramid}
\alias{lkp_pyramid}
\title{Age Pyramid}
\usage{
lkp_pyramid(
  df,
  x,
  y,
  f1,
  f2,
  fill,
  ...,
  y_current = 2018,
  n_breaks_y = NULL,
  facet_type = "wrap",
  scales = "free_x",
  nrow = NULL,
  ncol = NULL
)
}
\arguments{
\item{df}{A data.frame containing the data to represent}

\item{x}{The x axis variable for the histogram}

\item{y}{Count Variable}

\item{f1}{First optional variable for facet display}

\item{f2}{Second optional variable for facet display}

\item{fill}{A variable for which to show the breakdown in each histogram bar}

\item{...}{Additional parameters passed to the \code{"ll_pyramid"} function}

\item{y_current}{Current year, required for secondary scale}

\item{n_breaks_y}{Desired number of breaks for the y axis. Optional}

\item{facet_type}{Should \code{facet_wrap} or \code{facet_grid} be used for
facetting ?}

\item{scales}{Should scales be fixed (\code{"fixed"}, the default),
free (\code{"free"}), or free in one dimension (\code{"free_x"},
\code{"free_y"})?}

\item{nrow}{Number of rows and columns.}

\item{ncol}{Number of rows and columns.}
}
\value{
A ggplot object representing an age pyramid with one or several
facets
}
\description{
Age Pyramid
}
