% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{lkp_ridgeline}
\alias{lkp_ridgeline}
\title{Ridge Line Plot}
\usage{
lkp_ridgeline(
  df,
  x,
  y,
  f1,
  f2,
  ...,
  show_numbers,
  label_text,
  accuracy_text = NULL,
  vjust_text = -0.5,
  hjust_text = 0.5,
  label_x,
  label_y,
  accuracy_x = NULL,
  accuracy_y = NULL,
  n_breaks_x = NULL,
  n_breaks_y = NULL,
  facet_type = "wrap",
  scales = "fixed",
  nrow = NULL,
  ncol = NULL
)
}
\arguments{
\item{df}{A data.frame containing the data to represent}

\item{x}{The x axis variable for the histogram}

\item{y}{The y axis variable for the histogram}

\item{f1}{First optional variable for facet display}

\item{f2}{Second optional variable for facet display}

\item{...}{Additional parameters to be passed to \code{ll_ridgeline}.}

\item{show_numbers}{Should numbers be printed on top of the bars ? Optional}

\item{label_text}{Labeller for the printed numbers. Optional}

\item{accuracy_text}{Accuracy for the printed numbers. Optional}

\item{vjust_text}{Vertical adjustement for text}

\item{hjust_text}{Horizontal adjustement for text}

\item{label_x}{Labeller for the x axis. Optional}

\item{label_y}{Labeller for the y axis. Optional}

\item{accuracy_x}{Accuracy for the y axis. Optional}

\item{accuracy_y}{Accuracy for the y axis. Optional}

\item{n_breaks_x}{Desired number of breaks for the x axis. Optional}

\item{n_breaks_y}{Desired number of breaks for the y axis. Optional}

\item{facet_type}{Should \code{facet_wrap} or \code{facet_grid} be used for
facetting ?}

\item{scales}{Should scales be fixed (\code{"fixed"}, the default),
free (\code{"free"}), or free in one dimension (\code{"free_x"},
\code{"free_y"})?}

\item{nrow}{Number of rows and columns.}

\item{ncol}{Number of rows and columns.}
}
\value{
A ggplot object representing a ridgeline plot with one or several
facets
}
\description{
Ridge Line Plot
}
