% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_LinkPact}
\alias{theme_LinkPact}
\title{LinkPact ggplot theme}
\usage{
theme_LinkPact(
  base_size = 12,
  base_family = "serif",
  base_line_size = base_size/24,
  base_rect_size = base_size/24
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\value{
A complete ggplot2 theme.
}
\description{
LinkPact ggplot theme
}
\examples{
library(ggplot2)

theme_LinkPact() |> theme_set()

mtcars2 <- within(mtcars, {
  vs <- factor(vs, labels = c("V-shaped", "Straight"))
  am <- factor(am, labels = c("Automatic", "Manual"))
  cyl  <- factor(cyl)
  gear <- factor(gear)
})

(p1 <- ggplot(mtcars2) +
    geom_point(aes(x = wt, y = mpg, colour = gear)) +
    labs(title = "Fuel economy declines as weight increases",
         subtitle = "(1973-74)",
         caption = "Data from the 1974 Motor Trend US magazine.",
         x = "Weight (1000 lbs)",
         y = "Fuel economy (mpg)",
         colour = "Gears"))

p1 + facet_grid(vs ~ am)
}
